Pundit/Mani: In the UL the business customer creates the conference and modifies it. "Conference Owner" is not there in the UL. Updated to "Business Customer"
Pundit/Mani:
Currently the seats are locked using a hardcoded value, its not a value set by the conference owner. Using the term "cancels" the reservation leads to ambiguity in the content and code in this particular scenario where the seats are held/ blocked temporarily - cancel can also mean canceling a reservation user has paid for. Suggest changing this.
Changed terminology, also added note to check later how the timeout is specified.

Pundit/Peter Richie:
With a basis on DDD and the shear amount of interpretation over prescription; I think the use of Bounded Context in this guidance needs to be covered before being introduced here.If people don't understand how BC is used here; their foundation for understanding much of what follows is compromised.
Completed section "DEfinitions used in this chapter" and added links to more detailed information in the reference guide.