Pundit/Mani: Business Customer as per UL , not conference owner. - Updated.
Pundit/Mani:
Either change Conference Owner to Business Customer or fix UL.
- Updated.

Pundit/Mani:
MakeRegistration is not a command which has been implemented yet. Suggest providing an example which exists in the code.
- Changed to MakeSeatReservation.

Pundit/PeterRitchie:
A little unclear here. A task could also be a query. GetCustomerName could be a "command" for example. My interpretation of a command is that it's an action; something that changes state.
- Clarified this paragraph.

Pundit/Mani:
Any specfic why it should not have a return value and asynchronous? What was the motivation behind this design decision? Could explain his.
- The Deep Dive chapter (link in previous paragraph) discusses this in more detail.

Pundit/PeterRitchie:
"once and only once by a single recipient" could be inferred that there could be more than one recipient of a command but each recipient only handles the command once.
- Rephrased.

Pundit/Mani:
There is an order view model generator which subscribes to some events in the implementation , also there is the RegistrationProcessSagaRouter. There are different subscribers. The statement does not reflect the state of the implementation accurately.
- This was the state at this point in the journey - the OrderViewModelGenerator did not exist at this point in time. Added note to beginning of chapter.

Pundit/PeterRitchie:
Getting into "Aggregates", "publishing", "event bus", etc. without detailing these things yet. Most people looking for guidance on this topic aren't going to be knowledgeable about these things.
- These terms are covered in the Reference Guide - we don't want to clutter up the Journey guidance with too much detail. Added comment about the Reference Guide.

Pundit/Mani:
We are introducing the term event sourcing here without defining it , or providing a pointer to where it is defined.
- Added link to ES chapter in the Reference Guide

Pundit/Mani:
Not what is meant by X and Y here. This may have been a part of the conversation but it does not reflect the current implementation.
- Clarified language.

Pundit/Mani:
The diagram should be updated to use the term workflow? or the orignial term saga should be retained ... since its part of the journey and thats what it was called initially?
- Updated to reflect text.

Pundit/Mani:
Code snippet is not in sync with current implementation: private ConferenceDescriptionDTO GetConference(string conferenceCode) { var repo = this.repositoryFactory(); using (repo as IDisposable) { return repo.Query<ConferenceDescriptionDTO>() .Where(dto => dto.Code == conferenceCode) .FirstOrDefault(); } }
- This reflects the code at this point in the journey.

Pundit/Mani:
Code snippet needs to be updated
- This reflects the code at this point in the journey.