Included Pundit Feedback - also created some issues - see notes below:
Journey 00
Up front I think it should identify what the reader might get out of reading both documents and the reference implementation. Hook them in the first paragraph.or they might not read through all the material. Kellabyte.

Added Todo comment for missing introductory section.

I don't think a BC is a "subsystem" as described in this paragraph and alluded to in most of the Guidance. peterrichie

Modified language here, also added an issue (#56) to revisit this in the whole doc set.

Journey 01
Several comments from kellabyte. These relate to incomplete content - added comments to doc.
Also positive feedback on personae.

Journey 02
Link to the definition of bounded context in the CQRS context reference guide. BC's can communicate not only by events. Events are common with people doing CQRS+ES but it's not the only option. This text says "may only communicate with another bounded context by raising events". That might mislead people. Also link to what ubiquitous language means. - Kellabyte

Modified text and added callout, highlighted ubiquitious language.

Is this the context map? If so you should probably label it that and link to the definition of what a context map is. - Kellabyte

Added label and a comment to include a definition in the Ref Guide.

Link to a definition of domain in the CQRS context reference guide. - Kellabyte

There is already a link to the relevant chapter in the Ref Guide. We may later do a pass to more fully cross-reference terms like this.

Journey 03
"Orders and Reservations bounded context", I did't see a reservations bounded context on the previous page, I saw Orders and Registrations. Are we missing a BC in the context map? Later on it refers to the "Registrations bounded context". The uses of the BC naming isn't consistent with the context map on the previous page. -kellabyte

This naming issue is already addressed in the most recent version of the docs.

I would include the persona image and a short description in parenthesis the role this persona plays. It'll be hard to bounce back and forth reminding myself who is who. Something like: [image] Bharath (CQRS expert): "..." - kellabyte

This will happen as part of the production process for the docs.

There's something wrong with the formatting here. - PauloMorgado
Fixed.

Put arrows showing what buttons navigate to what screen. It's not obvious at first glance. - kellabyte

Created issue #57 to address this.

Use persona's here please. You created a great set of persona's and its a perfect opportunity to use them in these types of interactions. - kellabyte

We are considering this.

Only 2letter acronyms have both letters capitalized. Acronyms with 3 and more letters must be treated as words: ConferenceDto - PauloMorgado
I'm following what's in the code here.