Pundit Comments:
tibor19 - In most of todays projects, there is no architect role. I would call Janas role Technical Lead.

Although true, I'd prefer to keep the term Architect to help clarify the differences between Jana and Markus

manikrish - As of now- there is no registration required for admins.

True, but I think this is something that would be added if we had time - it's a valid requirement.