Pundit Comments: manikrish- Why send commands from server side ? This makes the logic a bit convoluted .. commands raising events which in turn raise commands on server side which can further raise events. As a part of the journey should explain why? this makes tracing things in code difficult, should add info on how the team had discussions with advisory to solve this problem etc.
Added link to Chapter 4.

manikrish - For v1 it is implemented as async but as a defintion this is misleading. https://groups.google.com/forum/#!msg/dddcqr

This is a working definition for this chapter - added link to chapter 6 where this is revised.

manikrish - Seat Ordering is not part of the UL

Changed section heading to Creating Orders

manikrish - No discounts in V1 - first stopping point.
manikrish - Waitlisting is not in V1 / first stopping point.

Changed to add future reference.