Pundit Comments:
manikrish - "An Order may have one or more Registrations associated with it." Looking at the implementation - An order can have multiple seat types associated with it, not registrations.

Looks like this should be Seat Assignment, not Registration - changed

manikrish - The ORder aggregate did have these states in the older code, not anymore

This chapter discusses the older code. No change.

manikrish - The purpose of the envelope is to wrap the message along with any properties/headers to be set on the message like the time delay. Could specify that here.

Already specified in next sentence.

manikrish - For v1 release :There is a separate topic for commands and events There is only one subscription each for these topics "all"The "routing" of events to handlers is done by the eventprocessor. We dont have a subscription for a set of handlers will use yet. saying that "Each subscription is associated with a particular handler type so that events reach all their subscribers" is incorrect.

Updated diagram and text to reflect V1 release.

manikrish - Add pointer to idempotency here. Also the behaviour in v1 was ... recieve, process, delete, should clarify this.

Updates to text.

manikrish - See previous comment on this. There is only 1 subscription right now for all the event handlers. There is a processor which sits in between and routes the events to the handlers which can handle those events.

Re-written this section to reflect actual implementation.

manikrish - one subscription per event type? topic per event type seems overkill, typo?

Typo - fixed.