Pundit Comments:
codingoutloud - Phrasing makes it sound like there is a single Orders and Registrations BC, not two. (OKAY - I see now there is only one - my oversight - and can't delete a pundit comment it seems. Speaking of Pundit - it should support ESC key or a close button so it does not interfere with my reading. It should not be in the way until I want it.)

Added issue to Pundit in Github

codingoutloud - he RDBMS object in the graph spells "Denormalized" w/o hyphen, though the text outside the graphic uses a hyphen ("De-normalized").

Standardized on "denormalized" throughout.

codingoutloud - Re: "The read-side should be optimized for fast reads" -- this deserves more than a passing reference - esp if true intent is to convey it as an explicit design goal, critical for UX responsiveness and application scalability. Further, may be more than "fast" but also low resource consumption.

Added a callout.

codingoutloud - "The team decided to adopt the second approach." would be more valuable if it included a reason: "The team decided to adopt the second approach because in this context the predictability and intent-revealing clarity of FindAll & Get were more important than the flexibility offered with IQueryable." (I don't know if that is the *real* reason, just giving possible example.)

Expanded text in line with recommendation.

manikrish - The snippet below is a bit confusing. A couple of paragraphs earlier we said the team selected using DAO approach. The snippet still shows IQueryable approach looks like.

Updated snippet