Pundit comments - Chapter 05
manikrish - InvoiceProcessorPaymentDetails not in v1

Text states this.

manikrish - What do we mean by element here, does it map to a code artifact/ project. Its a bit unclear now for example Order is an aggregate , but the Commands like RegisterToConference is recieved by the OrdercommandHandler and not the aggregate itself.

Added a callout explaining that handlers are omitted for reason of clarity - we are most interested in which commands go to which aggregates.

manikrish - The handlers would be a part of the BC wouldnt they? this information is missing in the diagram.

Added a callout explaining that handlers are omitted for reason of clarity - we are most interested in which commands go to which aggregates.

manikrish - InvoiceProcessorPaymentDetails not in v1