Pundit Comments:
jorgefioranelli - Wondering if we should explain why we have a Find method when Query can also filter by id. I believe it is an implementation detail, from my understanding Find is better because it searches first in memory and then in the database, while Query searches in the database only (at least in EF).

- I think this is slightly outside the scope - no change.

jorgefioranelli - I believe it is "read-model" instead of "write-model" here.

- Fixed

jorgefioranelli - Typo: informationa

- Fixed

jorgefioranelli - Another advantage of using this approach over IQueryable is that all the conditions are centralized in the same place. So if you need to add a new filter that affects more than one query, you don't need to check all the code that could do those queries.

- Added additional bullet point