Pundit feedback
manikrish: As general guidance- You can always create a worker role which completes the migration and provides a notification when its done, deploy the role only for the duration of the migration. That would be the preferred approach. Creating a worker role , then Remote deskop into box , running console seem to be extra steps- its just an implementation detail specific to the app and should not be taken as general guidance. The worker role could do the same thing the Program.cs main method for console app does with an addional step to send a notification via e-mail ( or you could check the logs periodically) to see if it migration completed.
Modified Poe's comment.

manikrish: The documentation on this is limited, however from one source which looks reliable -this does not seem to be true when using async send methods which we do. It does not look like we have solved this problem completely, should remove this as this misleads people who want to use the technology and use async. http://geekswithblogs.net/asmith/archive/2012/04/10/149275.aspx
Added issue to track this in GitHub

manikrish: See earlier comment on async send we do and sessions. This section potentially needs to be revisited.
Added issue in GitHub