Pundit feedback
gmelnik: "... attendees planning to come to a conference". Attend twice in the same place feels a bit tautological. :)
Done

gmelnik: "... the maximum number of seat*s*"
Done

gmelnik: not they, we
Done

gmelnik: "... in a later stage of the journey" => "... in some future journey" We need to mention that this is not implemented in the release, but members of the community are working on these features and we encourage the reader to check our main web site (cqrsjourney.github.com) for possible out-of-band releases and updates.
Done

gmelnik: ... as a result of a command, for example <b>OrderConfirmed</b>.
Done

gmelnik Might need to rename this to be consistent with the existing Process Manager patterns (from EIP book by G.Hohpe). Greg Young also seems to favour this term
Verifying this with the team later.

gmelnik: Not 'they', but 'we'
Done

gmelnik: Noticed an inconsistency with casing. This can be fixed during copy-edit but we need to advise Nancy on our preference. We either capitalize the persona name or not. We can also italicize it (but not capitalize).
Made a global change to captitalize Registrant, Business Customer, Attendee

gmelnik: CarlosPersona: We intentionally implemented a mechanism of returning to the previously submitted orders via the record locator mechanism. This eliminates an often annoying requirement for users to create an account in the system and sign in prior to be able to evaluate its usefulness. Our customers are adamant about this.
Added

gmelnik: Is this still true? Only 4 states? What about OrderUpdated? OrderExpired? I would map these states directly to our events (in \Registration.Contracts\Events)
In fact orders no longer have these explicit states - now changed to reflect this simplification.

gmelnik: A seat doesn't represent a concrete seat in the conference hall (like row 10, seat 42). It's just a right to be admitted to a certain conference session/event.
Changed

gmelnik: Insert "Each term was formulated through active discussions between the development team and the domain expert."The following is a sample conversation between...
Done

gmelnik: for <i>n</i> minutes
Done

gmelnik: ...the *ASP.NET* MVC...
Done

gmelnik: they => we
Done

gmelnik: and also, wouldn't there be a security concern if the only place that validation took place was the UI? because then, it would be easy to hack it to proceed with registration even if no seats are available, pay and then complain to the conference organizer showing a seemingly valid registration.
We are talking about the MVC controller, not the browser. Added a comment from Jana.

gmelnik: BethPersona: Undo is just one of many compensating actions that can be used in real life. Those compensating actions could even be outside of the system implementation and involve human actors (for example, Contoso clerks or the Business Customer calls the Registrant and tells them that some error was made and they should ignore the last confirmation email received from Contoso System).
Done

gmelnik: "consistency boundary" - new term introduced for the first time in the guide. Either need a reference or a quick term definition.
Done

gmelnik: You can download a copy of the code from the Download center (include fwlink) or check the evolution of the code from the repository on github:...
Done

gmelnik: No more Display action in this controller. But I guess it's ok since that was just an interim impl-n and we did make a disclaimer that the code in the repo may not match the code snippets in the certain parts of the book.
Agreed - no change made.

gmelnik: Any reference to the Post-Redirect-Get pattern?
Added Wikipedia link.

gmelnik: I've seen us using the term handler in the context of event handling.
Fixed this

gmelnik: Service Bus is not *basic*. Windows Azure queues are. I suggest removing basic from this sentence.
Done

gmelnik: (see Chapter XXX)
Added link

gmelnik: Why don't we discuss why using the Azure queues (not SB queues) are inadequate in this situation?
Updated Jana comment to explain this.

gmelnik: Impact on Testing
Done

gmelnik2h ago
See also the insightful point-counterpoint debate on the value of mocks b/w Steve Freeman, Nat Pryce and Joshua Kerievsky here (http://www.computer.org/portal/web/csdl/doi/10.1109/MS.2007.84)
Added link

gmelnik: "... coupled to the data store implementation (which is a smell):"
Done