Pundit comments
gmelnik: Also, what about the buffer that we've added to the actual timeout? to allow any in-flight transactions to finish (those that have started the payment step within the timeframe but ended payment when the time was up - the business decision was still to keep those transactions as valid as opposed to refunding the payments if they came a few mins late).
This is covered later.

gmelnik: Nice, I see this has been updated to reflect only active BCs.We can say something along the lines that "even with the seeming simplicity of the context map, the implementations of these BCs and more importantly their interactions are quite sophisticated and there're many controversies that were spun off these and many lessons learnt captured.
Done.

gmelnik: We need a small section or a note on the 3 bounded contexts that didn't make into the release: 1) Discounts BC, 2) Disconnected conf mgmt client, 3) Submissions & Schedule Management (implemented in Node.js). We don't have to dive into them, but as a teaser I think it's ok to mention them and to encourage community participation and future out-of-band releases.