Pundit feedback
gmelnik: Agreed to call it Process Manager. Pls rename. Also, make sure a proper reference to Gregor Hohpe's EIP book (pp.312-321) is made.
Done

gmelnik: Add CarlosPersona callout: From the business perspective it was important to be user-friendly. We don't want to block or unnecessarily burden any user who is trying to register for a conference. Thus, no requirement of having an account created in the system prior to registration, especially since most of the user's information will be entered in a standard checkout process anyway.
Done

gmelnik: A diagram here would really help grasp the architecture.
Done

gmelnik: Add JanaPersona: A normalized database schema can fail to provide adequate system response time due to excessive table join operations.Despite advances in relational databases, a JOIN is still a very expensive operation compared to a single-table read.
Done

gmelnik: This is the first time ViewRepository is mentioned. Should we describe what it's for and reference the Repository pattern. This can be done as a JanaPersona quote: The Repository pattern mediates between the domain and data mapping layers using a collection-like interface for accessing domain objects. (for more info see Fowler, P of EAA, online: http://martinfowler.com/eaaCatalog/repository.html )

gmelnik: However, if the dev team is cognizant of this downside and is committed to task-based UIs, this should not be an issue.
Done

gmelnik: MVC 4. Or simply say ASP.NET MVC.
ASP.NET MVC.

gmelnik: MarkusPersona: The Transient Fault Handling Application Block from Microsoft patterns & practices is designed to make it easier to implement consistent retry behavior for any transient faults. It comes with a set of built-in detection strategies for SQL Azure, Windows Azure Storage, Caching and Service Bus and also allows you to define your own strategies. Similarly, it comes with a set of handy built-in retry strategies and supports custom ones. For more info, see http://msdn.microsoft.com/en-us/library/hh680934(v=pandp.50)
Done

gmelnik: This doesn't always mean automated handling and recovery. Sometimes, manual resolution of a business failure might be sufficient.
Done

gmelnik: I don't know if this is too late now. But it would have been great if we could point to a check-in on github to get the version of the code that corresponded to this stop in the journey.
Too late here, but we have tags for the releases I believe.

gmelnik: assigning Attendees to seats (not registering). Don't want to confuse two actions.
Done

gmelnik: Is the buffer still 5 mins? The last time I spoke with Julian, he mentioned something like extra 15 mins to allow transactions that are already in the paying state to complete. Pls double check.
This was the buffer time used at this stage in the journey.

gmelnik: Unless we say that intially, at this stage of the journey, MVC3 was used and later the whole project was updated to MVC4 to improve installation experience of the RI and fewer dependencies.
Changed to just MVC.

gmelnik: Domain expert also paired with one of the testers on specifying core acceptance tests.
Done

gmelnik: replace "in the language" with "using SpecFlow notation".
Done

gmelnik: ... with members of the p&p team...
Done

gmelnik: remove "into a a powerful way"
Done