Pundit feedback
gmelnik: Add "Eventual consistency" definition here.Eventual consistency = a consistency model when after an update to a data object, the storage system does not guarantee that subsequent accesses to that object will return the updated value. However, the storage system guarantees that if no new updates are made to the object during a sufficiently long period of time, eventually all accesses can be expected to return the last updated value.
Done

gmelnik: "a Registrant may request..." is redundant.
Done

gmelnik: Insert "availability" in "consistency, availability, reliability..."
Done

gmelnik: Incorporate a quote from Greg Young:"I would like to state once and for all that CQRS does not require a task based UI. We could apply CQRS to a CRUD based interface (though things like creating separated data models would be much harder).There is however one thing that does really require a task based UI… That is Domain Driven Design."Source: http://codebetter.com/gregyoung/2010/02/16/cqrs-task-based-uis-event-sourcing-agh/
Done

gmelnik: Dominic, please see if you want to incorporate any of CRUD thinking from Ian Cooper's blog post: http://codebetter.com/iancooper/2011/07/15/why-crud-might-be-what-they-want-but-may-not-be-what-they-need/ or at least reference it. I think the discussion and especially the comments there are really good.
Added a reference.

gmelnik: .. the long-run when the diameter of the network increases, since both the response time and the probability of failure become higher.
Done

gmelnik: JanaPersona: This choice may change depending on the state of your system. Consider an overbooking scenario. The autonomy strategy may optimize for the normal case (lots of conference seats are still available), but as a particular conference is getting full, the system may need to become more conservative and favour authority (to use the latest info on seat availability).
Done

gmelnik: extra "i' - remove
Done

gmelnik: Additionally, the system introduces an additional, intrinsic buffer (which is not surfaced via the countdown clock) that allows a little extra time to complete the payment processing.
Done

gmelnik: Also, add another small section "Arguments against task-based UIs" (again, this is based on Avi's post): - Some say Task based UIs are less flexible the CRUD, with CRUD we could change the business process at any time and not change the software.Well, they are right!But… What will provide more value to the business?Is the ability to change the business process in one second worth more than the time and mistakes saved by guiding the user through the process?What about reports? Could we analyze what’s going on in the business just by knowing that data has changed?"Also, bring on the YAGNI thing. Often, people design their systems with the capabilities to extend and modify the processes and at the end hardly anyone uses those features. So, simplicity for the win!
Some of this was already in the ref guide - added links.

gmelnik: Do you think we could have used the output from Josh's MIL here? Or at least we can refer the reader to the section in the previous chapter that talks about MIL as a possible way to grok the architecture.
MIL seems to be broken at the moment, I emailed Josh to see if there's a quick resolution.