Pundit feedback
fsimonazzi: It was asynchronous in this version, but it became synchronous in v3, at least for some commands.
This is already described in journey 07 - no change.

fsimonazzi: The conf management system publishes events that do not originate in ARs
Modified.

fsimonazzi: Does this mean that we don't consider read model generators to be part of the BC?
Modified.

fsimonazzi: Is the message or the handling that is idempotent? I can see how some messages lend themselves better to idempotent processing.
Modified.

fsimonazzi: The web admin talks directly to the db as well, for reads and writes, but it has its own set of tables.Read models generators might need to read from the db before pushing updates.
Agreed, but want to keep diagram simple/high level here.

fsimonazzi: Even more. Now that each handler has its own subscription it could be that each handler has a different processing time for each event.
Discussed in next chapter.

fsimonazzi: Yes, but events are ordered for a specific AR. There is no global order for events. At least with the implementation we're using.
Clarified

fsimonazzi: typo exsiting
Fixed

fsimonazzi: This is not entirely accurate. The aggregate roots don't send their events themselves.The repositories will, when saving an AR, persist its events to storage and then a different process (the EventStoreBusPublisher) will asynchronously pick up the events from storage and publish them in order and waiting for the current event from an AR to be published before publishing the next one.
Fixed

fsimonazzi: The question here would be why bother sending async if you're waiting for the operation to complete anyway, thus blocking the thread.
Asked for clarification.

fsimonazzi: We could compare this to the choice of partition and row ids in the event store. Even though the schema is pretty much the same, the choice of ids reflects their intended purposes.
Done

fsimonazzi: This is a best effort approach, and will generate events representing the current state. It cannot generate intermediate events. Also, these synthetic events are assumed to have happened all before the public site started operations.
Done

fsimonazzi: And the worker role does not process messages.
Done

fsimonazzi: Each role has its own MaintenanceMode property. This is relevant for the migration to V3.
Done

fsimonazzi: Have we described the migration to a new version of the event store azure table with additional fields?
Covered earlier.

fsimonazzi: "The following sections" start this paragraph and the one before the previous paragraph. It's a bit repetitive.
Removed