Pundit feedback
Journey_02
jelster: Making the conscious decision that Conf. Mgmt would be a CRUD system was the right choice. In retrospect, it demonstrated how existing systems can be folded into the CQRS model as part of a greater whole.
No change.

Journey_03
jelster: Could use some MIL here
MIL not introduced until Chapter 4 - no change.

jelster: this is good. I think it may be useful to present this earlier in the material, even if in a summary or abbreviated form. The earlier people are able to picture it, the more the explanations will make sense
This is the first time we have talked about the infrastructure - no change.

jelster: everything" may be a bit vague for folks unfamiliar with terminology. Perhaps "...boundary within which state changes are assumed to propagate instantaneously" would be more helpful?
Changed definition.

Epilogue
ploeh: Isn't that what "Enterprise Integration Patterns" ( http://www.amazon.com/gp/product/0321200683/ref=as_li_ss_tl?ie=UTF8&camp=1789&creative=390957&creativeASIN=0321200683&linkCode=as2&tag=ploeh-20&l=as2&o=1&a=0321200683 ) already provides?
The comment was about communicating with non-technical people - EIP does not quite fit the bill, MIL does help. No change.

ploeh: So the key take-away here is really that all of this infrastructure is what Eric Evans calls 'generic sub-domains'. These are often better dealt with by leveraging existing software (e.g. NServiceBus, Rebus, etc.) instead of giving in to one's NIH syndrome.
This point is emphasized later in the chapter - no change.

ploeh: I think this is an important lesson that bears further elaboration.CQRS has lots of moving parts, and that scares people. There has to be a significant benefit from adding all of those moving parts, or you're not going to convince many people to adopt it.One benefit is buried in this paragraph.What's best?- Spending a long time finding an issue, and then fixing it quickly- Finding an issue quickly, and the spend a long time fixing itThe reason the first option is better is that not only can you fix it quickly, but you can also make sure that it stays fixed. This is because the separation of concerns provides a decoupled architecture that makes it easy to reproduce issues as unit tests.
Extended this paragraph.

ploeh: Well, *if* TDD is done bottom-up, it's not a good match for exploration.However, I'm not sure there's evidence for the statement that "TDD tends to be used in a bottom-up approach".That's certainly not how I write application code, and neither is it the style recommended by GOOS ( http://www.amazon.com/gp/product/0321503627/ref=as_li_ss_tl?ie=UTF8&camp=1789&creative=390957&creativeASIN=0321503627&linkCode=as2&tag=ploeh-20&l=as2&o=1&a=0321503627 )Another thing entirely is that this paragraph seems to contradict itself. First it indicates that TDD makes it hard, and then it concludes by discussing the help that TDD can give.
Agreed with team to delete this paragraph.